requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
censys==1.1.1
dnsdb==0.2.5
dnspython==2.0.0
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
ipwhois==1.2.0
itsdangerous==1.1.0
netaddr==0.8.0
pybinaryedge==0.5
PyJWT==1.5.3
pymongo==3.11.3
python-dotenv==0.17.0
pyvis==0.1.9
rapidfuzz==1.4.1
redis==3.5.3
requests==2.25.1
rq==1.8.0
shodan==1.25.0
tldextract==3.1.0
urllib3==1.26.4
uWSGI==2.0.19.1
WTForms==2.3.3
XlsxWriter==1.4.0
